requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0  # For Parquet support
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Web scraping and browser automation
selenium>=4.15.0
undetected-chromedriver>=3.5.0
fake-useragent>=1.4.0

# Text processing and NLP
scikit-learn>=1.3.0
nltk>=3.8.0
textblob>=0.17.0
regex>=2023.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Async and concurrency
aiohttp>=3.8.0
# asyncio
# concurrent.futures

# Data validation and configuration
pydantic>=2.4.0
pyyaml>=6.0.0

# Logging and monitoring
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.9.0
flake8>=6.0.0
mypy>=1.6.0

# Time zone handling
pytz>=2023.3
python-dateutil>=2.8.0

# Rate limiting
ratelimit>=2.2.0

# Hash functions for deduplication
xxhash>=3.4.0


#emoji modules
emoji>=2.2.0